requirements.txt
Pillow==9.4.0
pyzbar==0.1.9
